requirements.txt
# Core CV & API
opencv-python>=4.8.0
mediapipe>=0.10.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# AWS & Cloud
boto3>=1.34.0
botocore>=1.34.0

# MLOps
mlflow>=2.9.0
dvc>=3.30.0

# Testing & Quality
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0
moto[ec2]>=4.2.0

# Utilities
python-dotenv>=1.0.0
streamlit>=1.28.0  # Optional frontend
requests>=2.31.0